requirements.txt
flask
transformers
scikit-learn
torch
python-dateutil